[Java] Enabled position tracking in the presence of termIds rolling over to be negative and optimised imports.

+++ aeron-client/src/main/java/uk/co/real_logic/aeron/Subscription.java
@@ -92,7 +92,7 @@ public class Subscription implements AutoCloseable
     }
 
     public void onTermBuffersMapped(final long sessionId,
-                                    final long termId,
+                                    final int termId,
                                     final LogReader[] logReaders,
                                     final PositionReporter positionReporter,
                                     final ManagedBuffer[] managedBuffers)
