+++ aeron-driver/src/main/java/uk/co/real_logic/aeron/driver/DriverSubscriptionDispatcher.java
@@ -99,7 +99,7 @@ public class DriverSubscriptionDispatcher
         if (null != subscription)
         {
             final long sessionId = header.sessionId();
-            final long termId = header.termId();
+            final int termId = header.termId();
             final DriverConnectedSubscription connectedSubscription = subscription.getConnectedSubscription(sessionId);
 
             if (null != connectedSubscription)
