+++ aeron-driver/src/main/java/uk/co/real_logic/aeron/driver/SenderControlStrategy.java
@@ -24,7 +24,7 @@ public interface SenderControlStrategy
      *
      * @return the calculated rightEdgeOfWindow
      */
-    long onStatusMessage(final long termId,
+    long onStatusMessage(final int termId,
                          final long highestContiguousSequenceNumber,
                          final long receiverWindow,
                          final InetSocketAddress address);
@@ -36,5 +36,5 @@ public interface SenderControlStrategy
      * @param termBufferCapacity to use as the size of each term buffer
      * @return limit to which the position can go.
      */
-    long initialPositionLimit(final long initialTermId, final int termBufferCapacity);
+    long initialPositionLimit(final int initialTermId, final int termBufferCapacity);
 }
