+++ aeron-driver/src/test/java/uk/co/real_logic/aeron/driver/LossHandlerTest.java
@@ -56,7 +56,7 @@ public class LossHandlerTest
     private static final int ALIGNED_FRAME_LENGTH = align(MESSAGE_LENGTH, FrameDescriptor.FRAME_ALIGNMENT);
     private static final long SESSION_ID = 0x5E55101DL;
     private static final long CHANNEL_ID = 0xC400EL;
-    private static final long TERM_ID = 0xEE81D;
+    private static final int TERM_ID = 0xEE81D;
 
     public static final StaticDelayGenerator delayGenerator =
         new StaticDelayGenerator(TimeUnit.MILLISECONDS.toNanos(20), false);
