+++ aeron-system-tests/src/test/java/uk/co/real_logic/aeron/PubUnicastTest.java
@@ -15,12 +15,18 @@
  */
 package uk.co.real_logic.aeron;
 
-import org.junit.*;
+import org.junit.After;
+import org.junit.Before;
+import org.junit.Ignore;
+import org.junit.Test;
 import uk.co.real_logic.aeron.common.BitUtil;
 import uk.co.real_logic.aeron.common.concurrent.AtomicBuffer;
 import uk.co.real_logic.aeron.common.concurrent.logbuffer.FrameDescriptor;
 import uk.co.real_logic.aeron.common.event.EventLogger;
-import uk.co.real_logic.aeron.common.protocol.*;
+import uk.co.real_logic.aeron.common.protocol.DataHeaderFlyweight;
+import uk.co.real_logic.aeron.common.protocol.HeaderFlyweight;
+import uk.co.real_logic.aeron.common.protocol.NakFlyweight;
+import uk.co.real_logic.aeron.common.protocol.StatusMessageFlyweight;
 import uk.co.real_logic.aeron.driver.MediaDriver;
 
 import java.net.InetSocketAddress;
@@ -28,6 +34,7 @@ import java.nio.ByteBuffer;
 import java.nio.channels.DatagramChannel;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
+import java.util.concurrent.atomic.AtomicInteger;
 import java.util.concurrent.atomic.AtomicLong;
 
 import static org.hamcrest.MatcherAssert.assertThat;
@@ -122,7 +129,7 @@ public class PubUnicastTest
             Thread.yield();
         }
 
-        final AtomicLong termId = new AtomicLong();
+        final AtomicInteger termId = new AtomicInteger();
         final AtomicLong receivedZeroLengthData = new AtomicLong();
         final AtomicLong receivedDataFrames = new AtomicLong();
 
@@ -185,7 +192,7 @@ public class PubUnicastTest
             Thread.yield();
         }
 
-        final AtomicLong termId = new AtomicLong();
+        final AtomicInteger termId = new AtomicInteger();
         final AtomicLong receivedZeroLengthData = new AtomicLong();
         final AtomicLong receivedDataFrames = new AtomicLong();
 
@@ -260,7 +267,7 @@ public class PubUnicastTest
         EventLogger.logInvocation();
     }
 
-    private void sendSM(final long termId) throws Exception
+    private void sendSM(final int termId) throws Exception
     {
         final ByteBuffer smBuffer = ByteBuffer.allocate(StatusMessageFlyweight.HEADER_LENGTH);
         statusMessage.wrap(new AtomicBuffer(smBuffer), 0);
@@ -282,7 +289,7 @@ public class PubUnicastTest
         assertThat(bytesSent, is(StatusMessageFlyweight.HEADER_LENGTH));
     }
 
-    private void sendNak(final long termId, final long termOffset, final long length) throws Exception
+    private void sendNak(final int termId, final long termOffset, final long length) throws Exception
     {
         final ByteBuffer nakBuffer = ByteBuffer.allocate(NakFlyweight.HEADER_LENGTH);
         nakHeader.wrap(new AtomicBuffer(nakBuffer), 0);
