+++ aeron-system-tests/src/test/java/uk/co/real_logic/aeron/SubUnicastTest.java
@@ -54,7 +54,7 @@ public class SubUnicastTest
     private static final String DESTINATION = "udp://" + HOST + ":" + PORT;
     private static final long CHANNEL_ID = 1L;
     private static final long SESSION_ID = 2L;
-    private static final long TERM_ID = 3L;
+    private static final int TERM_ID = 3;
     private static final byte[] PAYLOAD = "Payload goes here!".getBytes();
     private static final byte[] NO_PAYLOAD = {};
     private static final int FRAME_COUNT_LIMIT = Integer.MAX_VALUE;
